udict_inline: fix udict_pool_depth type

+++ lib/upipe/udict_inline.c
@@ -663,7 +663,7 @@ static void udict_inline_mgr_free(struct urefcount *urefcount)
  * (if set to -1, a default sensible value is used)
  * @return pointer to manager, or NULL in case of error
  */
-struct udict_mgr *udict_inline_mgr_alloc(unsigned int udict_pool_depth,
+struct udict_mgr *udict_inline_mgr_alloc(uint16_t udict_pool_depth,
                                          struct umem_mgr *umem_mgr,
                                          int min_size, int extra_size)
 {
